move selection-tooltip to common folder
