+++ packages/ember-rdfa-editor/src/components/_private/common/selection-tooltip.ts
